Open `FieldType.extractMessageType()` and `extractTypeName()`
Also:
 * Improve code layout.
