+++ api/src/main/kotlin/io/spine/protodata/ast/FieldTypeExts.kt
@@ -34,14 +34,9 @@ import io.spine.protodata.ast.FieldType.KindCase.LIST
 import io.spine.protodata.ast.FieldType.KindCase.MAP
 import io.spine.protodata.ast.FieldType.KindCase.MESSAGE
 import io.spine.protodata.ast.FieldType.KindCase.PRIMITIVE
+import io.spine.protodata.type.TypeSystem
 import io.spine.string.shortly
 import io.spine.string.simply
-
-/**
- * Tells if this type holds one value such as primitive, message, or enum item.
- */
-public val FieldType.isSingular: Boolean
-    get() = isMessage || isEnum || isPrimitive
 
 /**
  * Obtains a human-readable name of this field type.
@@ -54,6 +49,39 @@ public val FieldType.name: String
         LIST -> "repeated ${list.name}"
         MAP -> "map<${map.keyType.name}, ${map.valueType.name}"
         else -> kindCase.name
+    }
+
+/**
+ * Indicates if this field is `repeated`, but not a `map`.
+ */
+public val FieldType.isList: Boolean
+    get() = hasList()
+
+/**
+ * Indicates if this field type represents a message.
+ */
+public val FieldType.isMap: Boolean
+    get() = hasMap()
+
+/**
+ * Indicates if this type holds one value such as primitive, message, or enum item.
+ */
+public val FieldType.isSingular: Boolean
+    get() = isMessage || isEnum || isPrimitive
+
+/**
+ * Obtains a cardinality of this field type.
+ *
+ * @see Cardinality
+ */
+public val FieldType.cardinality: Cardinality
+    get() = when (kindCase) {
+        MESSAGE,
+        ENUMERATION,
+        PRIMITIVE -> CARDINALITY_SINGLE
+        LIST -> CARDINALITY_LIST
+        MAP -> CARDINALITY_MAP
+        else -> error("Unable to convert `$kindCase` to `Cardinality`.")
     }
 
 /**
@@ -73,28 +101,44 @@ public fun FieldType.toType(): Type = type {
 }
 
 /**
- * Tells if this field is `repeated`, but not a `map`.
- */
-public val FieldType.isList: Boolean
-    get() = hasList()
-
-/**
- * Tells if this field type represents a message.
- */
-public val FieldType.isMap: Boolean
-    get() = hasMap()
-
-/**
- * Obtains a cardinality of this field type.
+ * Obtains full message type information for singular fields, lists, or maps storing messages.
  *
- * @see Cardinality
+ * @param typeSystem The type system to be used for obtaining type information.
+ *
+ * @return the message type instance or `null` if this field type is not a message,
+ *   or if it does not refer to message being a list or a map.
  */
-public val FieldType.cardinality: Cardinality
-    get() = when (kindCase) {
-        MESSAGE,
-        ENUMERATION,
-        PRIMITIVE -> CARDINALITY_SINGLE
-        LIST -> CARDINALITY_LIST
-        MAP -> CARDINALITY_MAP
-        else -> error("Unable to convert `$kindCase` to `Cardinality`.")
-    }
+public fun FieldType.extractMessageType(typeSystem: TypeSystem): MessageType? = when {
+    isMessage -> message.toMessageType(typeSystem)
+    isList -> list.maybeMessageType(typeSystem)
+    isMap -> map.valueType.maybeMessageType(typeSystem)
+    else -> null
+}
+
+/**
+ * Optionally converts this type into [MessageType] if this type is a message.
+ */
+private fun Type.maybeMessageType(typeSystem: TypeSystem): MessageType? =
+    if (isMessage) toMessageType(typeSystem) else null
+
+/**
+ * Obtains a name of a field type if it is a message or an enum.
+ *
+ * @return the name of the message or enum type, or `null` otherwise.
+ */
+public fun FieldType.extractTypeName(): TypeName? = when {
+    isMessage -> message
+    isEnum -> enumeration
+    isList -> list.maybeTypeName()
+    isMap -> map.valueType.maybeTypeName()
+    else -> null
+}
+
+/**
+ * Obtains the name of this type for enums and messages, returning `null` otherwise.
+ */
+private fun Type.maybeTypeName(): TypeName? = when {
+    isMessage -> message
+    isEnum -> enumeration
+    else -> null
+}
